+++ packages/rlc-common/src/interfaces.ts
@@ -203,7 +203,7 @@ export interface PackageDetails {
   scopeName?: string;
   nameWithoutScope?: string;
   description?: string;
-  version: string;
+  version?: string;
 }
 export interface OperationParameter {
   operationGroup: string;
@@ -237,6 +237,7 @@ export interface ParameterMetadata {
 export interface OperationResponse {
   operationGroup: string;
   operationName: string;
+  path: string;
   responses: ResponseMetadata[];
 }
 export interface ResponseMetadata {