+++ packages/rlc-common/src/metadata/buildReadmeFile.ts
@@ -169,6 +169,7 @@ function createMetadata(model: RLCModel): Metadata | undefined {
   const clientClassName = getClientName(model);
   const serviceName = getServiceName(model);
   let apiRefUrlQueryParameter: string = "";
+  packageDetails.version = packageDetails.version ?? "1.0.0-beta.1";
   if (packageDetails?.version.includes("beta")) {
     apiRefUrlQueryParameter = "?view=azure-node-preview";
   }