+++ packages/typespec-ts/src/modular/buildClientContext.ts
@@ -3,31 +3,29 @@ import { getClientParameters } from "./helpers/clientHelpers.js";
 import { importCredential } from "./helpers/credentialHelpers.js";
 import { getClientName } from "./helpers/namingHelpers.js";
 import { Client, Parameter } from "./modularCodeModel.js";
+import { isRLCMultiEndpoint } from "../utils/clientUtils.js";
+import { SdkContext } from "@azure-tools/typespec-client-generator-core";
+import { getDocsFromDescription } from "./helpers/docsHelpers.js";
 
 /**
  * This function creates the file containing the modular client context
  */
 export function buildClientContext(
+  dpgContext: SdkContext,
   client: Client,
   project: Project,
-  srcPath: string = "src"
+  srcPath: string = "src",
+  subfolder: string = ""
 ): SourceFile {
   const { description, parameters } = client;
   const name = getClientName(client);
   const clientContextFile = project.createSourceFile(
-    `${srcPath}/src/api/${name}Context.ts`
+    `${srcPath}/src/${
+      subfolder && subfolder !== "" ? subfolder + "/" : ""
+    }/api/${name}Context.ts`
   );
 
-  clientContextFile.addImportDeclaration({
-    moduleSpecifier: "../rest/index.js",
-    namedImports: [`${client.name}Context`]
-  });
-
-  clientContextFile.addExportDeclaration({
-    moduleSpecifier: "../rest/index.js",
-    namedExports: [`${client.name}Context`]
-  });
-
+  let factoryFunction;
   clientContextFile.addImportDeclaration({
     moduleSpecifier: "@azure-rest/core-client",
     namedImports: ["ClientOptions"]
@@ -39,13 +37,47 @@ export function buildClientContext(
     extends: ["ClientOptions"]
   });
 
-  const factoryFunction = clientContextFile.addFunction({
-    docs: [description],
-    name: `create${name}`,
-    returnType: `${client.name}Context`,
-    parameters: getClientParameters(client),
-    isExported: true
-  });
+  if (isRLCMultiEndpoint(dpgContext)) {
+    clientContextFile.addImportDeclaration({
+      moduleSpecifier: `../../rest/${subfolder}/index.js`,
+      namedImports: [`Client`]
+    });
+
+    clientContextFile.addExportDeclaration({
+      moduleSpecifier: `../../rest/${subfolder}/index.js`,
+      namedExports: [`Client`]
+    });
+    factoryFunction = clientContextFile.addFunction({
+      docs: getDocsFromDescription(description),
+      name: `create${name}`,
+      returnType: `Client.${client.name}`,
+      parameters: getClientParameters(client),
+      isExported: true
+    });
+  } else {
+    const rlcClientName = client.rlcClientName;
+    clientContextFile.addImportDeclaration({
+      moduleSpecifier: `${
+        subfolder && subfolder !== "" ? "../" : ""
+      }../rest/index.js`,
+      namedImports: [`${rlcClientName}`]
+    });
+
+    clientContextFile.addExportDeclaration({
+      moduleSpecifier: `${
+        subfolder && subfolder !== "" ? "../" : ""
+      }../rest/index.js`,
+      namedExports: [`${rlcClientName}`]
+    });
+
+    factoryFunction = clientContextFile.addFunction({
+      docs: getDocsFromDescription(description),
+      name: `create${name}`,
+      returnType: `${rlcClientName}`,
+      parameters: getClientParameters(client),
+      isExported: true
+    });
+  }
 
   const credentialsParam = parameters.find(
     (p) => p.clientName === "credential"
@@ -55,32 +87,49 @@ export function buildClientContext(
     (p) => p.location === "endpointPath"
   );
 
-  let baseUrl: string | undefined = "endpoint";
   if (baseUrlParam) {
+    let baseUrl: string | undefined = "endpoint";
     baseUrl =
       baseUrlParam.type.type === "constant"
         ? baseUrlParam.type.value
         : baseUrlParam.clientName;
+    factoryFunction.addStatements([`const baseUrl = ${baseUrl}`]);
   }
 
-  factoryFunction.addStatements([`const baseUrl = ${baseUrl}`]);
-  let getClientStatement = `const clientContext = getClient(baseUrl, options)`;
+  let getClientStatement = `const clientContext = getClient(options)`;
+
+  if (baseUrlParam) {
+    getClientStatement = `const clientContext = getClient(baseUrl, options)`;
+  }
 
   // If the client needs credentials we need to pass those to getClient
   if (credentialsParam) {
     importCredential(credentialsParam.type, clientContextFile);
     addCredentialOptionsStatement(credentialsParam, factoryFunction);
-    getClientStatement = `const clientContext = getClient(baseUrl, credential, options)`;
+    getClientStatement = baseUrlParam
+      ? `const clientContext = getClient(baseUrl, credential, options)`
+      : `const clientContext = getClient(credential, options)`;
   }
 
   factoryFunction.addStatements([getClientStatement, "return clientContext;"]);
 
-  clientContextFile.addImportDeclarations([
-    {
-      moduleSpecifier: "../rest/index.js",
-      defaultImport: "getClient"
-    }
-  ]);
+  if (isRLCMultiEndpoint(dpgContext)) {
+    clientContextFile.addImportDeclarations([
+      {
+        moduleSpecifier: `../../rest/${subfolder}/index.js`,
+        namedImports: ["createClient as getClient"]
+      }
+    ]);
+  } else {
+    clientContextFile.addImportDeclarations([
+      {
+        moduleSpecifier: `${
+          subfolder && subfolder !== "" ? "../" : ""
+        }../rest/index.js`,
+        defaultImport: "getClient"
+      }
+    ]);
+  }
 
   clientContextFile.fixMissingImports(
     {},