+++ packages/typespec-ts/test/commands/cadl-ranch-list.ts
@@ -1,10 +1,14 @@
-export interface CadlRanchConfig {
+export interface TypeSpecRanchConfig {
   outputPath: string;
   inputPath: string;
   debug?: boolean;
 }
 
-export const cadls: CadlRanchConfig[] = [
+export const rlcTsps: TypeSpecRanchConfig[] = [
+  {
+    outputPath: "union-body",
+    inputPath: "union-body"
+  },
   {
     outputPath: "headers/clientRequestId",
     inputPath: "special-headers/client-request-id"
@@ -148,5 +152,12 @@ export const cadls: CadlRanchConfig[] = [
   {
     outputPath: "azure/core-traits",
     inputPath: "azure/core/traits"
+  }
+];
+
+export const modularTsps: TypeSpecRanchConfig[] = [
+  {
+    outputPath: "models/usage",
+    inputPath: "type/model/usage"
   }
 ];